% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beispiel.daten.R
\name{beispiel.daten}
\alias{beispiel.daten}
\title{Funktion zum herunterladen der Beispieldaten}
\usage{
beispiel.daten(pfad = tempdir(), unzip = TRUE)
}
\arguments{
\item{pfad}{Pfad zu dem Ordner in dem die Daten gespeichert werden sollen. Sollte der Ordner nicht existieren, wird er erstellt. Wenn kein Pfad angegeben wird, werden die Daten in einem temporären Ordner gespeichert.}

\item{unzip}{Wenn TRUE, werden die Daten entpackt. Standard ist TRUE.}
}
\description{
Die Funktion laedt die Beispieldaten aus dem Buch herunter. Die Daten werden in dem angegebenene Ordner gespeichert. Wenn das Paket installiert wurde, sind die Daten auch im Ordner 'extdata' zu finden in dem R Packet Ulmerbuch gespeichert..
}
