% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beispiel.pfad.R
\name{beispiel.pfad}
\alias{beispiel.pfad}
\title{Funktion zum Setzen des Pfades von Beispiel Daten}
\usage{
beispiel.pfad(pfad = NULL)
}
\arguments{
\item{pfad}{entweder NULL oder ein Pfad zu einem Ordner. Wenn NULL wird der Pfad zu den Beispiel Daten des Pakets gesetzt. Wenn ein Pfad angegeben wird, wird der Pfad zu diesem Ordner gesetzt.}
}
\description{
Diese Funktion hilft beim Laden der Beispiel Daten. Wenn das Paket "Ulmerbuch" installiert wurde befinden sich alle Beispiel Daten im Ordner "extdata". Diese Funktion setzt den Pfad zu den Beispiel Daten. Manchmal ist es auch wuenschenswert seine eigenen Daten zu verwenden dann ist eine Anpassung des Pfades notwendig. Wenn die Funktion ohne Parameter aufgerufen wird, wird der Pfad zu den Beispiel Daten gesetzt in 'extdata' gesetzt. Wenn eine Pfad angegeben wird, wird der Pfad zu diesem Ordner gesetzt. Die Namen der Files muessen dann im Beipsiel Code entsprechend der persoenlichen Daten angepasst werden.
}
