% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beispiel.R
\name{beispiel}
\alias{beispiel}
\title{Funktion zum Oeffnen von Beispiel code}
\usage{
beispiel(bsp = NULL)
}
\arguments{
\item{bsp}{Beispiel Nummber aus dem Buch. Bitte in Anfuehrungezeichen angeben. Zum Beispiel wird Beispiel 4.1 mittles `bsp="4.1"` aufgerufen.}
}
\description{
Die Funktion oeffnet den Beispiel Code als ein Rstudio file. Das Code file kannst du beliebig veraendern und Deinen Beduerfnissen anpassen. Bitte am besten das veraenderte Beispiel unter neuen Namen und an anderer Stelle abspeichern.
}
