\name{UlmerBuch-package}
\alias{UlmerBuch-package}
\alias{UlmerBuch}
\docType{package}
\title{
\packageTitle{UlmerBuch}
}
\description{
This is the accompanying package to run examples for the Ulmerbuch.
Check the website for details \url{https://github.com/green-striped-gecko/UlmerBuch}.
and one more sentence.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UlmerBuch}
\packageIndices{UlmerBuch}
}
\author{
\packageAuthor{UlmerBuch}
Maintainer: \packageMaintainer{UlmerBuch}
}
