% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste.beispiele.R
\name{liste.beispiele}
\alias{liste.beispiele}
\title{Liste aller Beispiele}
\usage{
liste.beispiele(links = TRUE)
}
\arguments{
\item{links}{Logischer Wert. Falls TRUE, wird eine interaktive Tabelle
mit anklickbaren Links zu PDF- und Excel-Dateien angezeigt.
Standard ist TRUE.}
}
\value{
Eine Tabelle mit den vorhandenen Beispielen und Dateiformaten.
}
\description{
Gibt eine Übersicht über alle im Paket vorhandenen Beispiele aus.
Optional werden anklickbare Links zu PDF- und Excel-Dateien angezeigt.
}
\examples{
liste.beispiele()
liste.beispiele(links = TRUE)

}
