% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste.beispiele.R
\name{liste.beispiele}
\alias{liste.beispiele}
\title{Funktion zum Auflisten aller Beispiele}
\usage{
liste.beispiele()
}
\description{
Gibt eine Lister aller vorhandener Beispiele des Ulmerbuchs aus. 
Einige Beispiele sind als pdf und in rmd (code der direkt ausgefuehrt werden kann) 
format vorhankden.
}
\examples{
liste.beispiele()
}
