\name{UlmerBuch-package}
\alias{UlmerBuch-package}
\alias{UlmerBuch}
\docType{package}
\title{
\packageTitle{UlmerBuch}
}
\description{
\packageDescription{UlmerBuch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UlmerBuch}
\packageIndices{UlmerBuch}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{UlmerBuch}

Maintainer: \packageMaintainer{UlmerBuch}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
